+++ yundun-bastionhost-20191209/core/Models/CreateOperationTicketRequest.cs
@@ -1,0 +1,93 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Yundun_bastionhost20191209.Models
+{
+    public class CreateOperationTicketRequest : TeaModel {
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// </summary>
+        [NameInMap("ApproveComment")]
+        [Validation(Required=false)]
+        public string ApproveComment { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>root</para>
+        /// </summary>
+        [NameInMap("AssetAccountName")]
+        [Validation(Required=false)]
+        public string AssetAccountName { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>2</para>
+        /// </summary>
+        [NameInMap("AssetId")]
+        [Validation(Required=false)]
+        public string AssetId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>1679393152</para>
+        /// </summary>
+        [NameInMap("EffectEndTime")]
+        [Validation(Required=false)]
+        public long? EffectEndTime { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>1685600242</para>
+        /// </summary>
+        [NameInMap("EffectStartTime")]
+        [Validation(Required=false)]
+        public long? EffectStartTime { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>bastionhost-cn-st220aw****</para>
+        /// </summary>
+        [NameInMap("InstanceId")]
+        [Validation(Required=false)]
+        public string InstanceId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>false</para>
+        /// </summary>
+        [NameInMap("IsOneTimeEffect")]
+        [Validation(Required=false)]
+        public bool? IsOneTimeEffect { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>SSH</para>
+        /// </summary>
+        [NameInMap("ProtocolName")]
+        [Validation(Required=false)]
+        public string ProtocolName { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>cn-hangzhou</para>
+        /// </summary>
+        [NameInMap("RegionId")]
+        [Validation(Required=false)]
+        public string RegionId { get; set; }
+
+    }
+
+}