+++ yundun-bastionhost-20191209/core/Models/CreateOperationTicketResponseBody.cs
@@ -1,0 +1,30 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Yundun_bastionhost20191209.Models
+{
+    public class CreateOperationTicketResponseBody : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>2</para>
+        /// </summary>
+        [NameInMap("OperationTicketId")]
+        [Validation(Required=false)]
+        public string OperationTicketId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>0ECCC399-4D35-48A7-8379-5C6180E66235</para>
+        /// </summary>
+        [NameInMap("RequestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+    }
+
+}