+++ yundun-bastionhost-20191209/core/Models/GetExportConfigJobResponseBody.cs
@@ -1,0 +1,62 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Yundun_bastionhost20191209.Models
+{
+    public class GetExportConfigJobResponseBody : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para><a href="https://bastionhost-cn-hangzhou-164***.oss-cn-hangzhou.aliyuncs.com/bastionhost-cn-2">https://bastionhost-cn-hangzhou-164***.oss-cn-hangzhou.aliyuncs.com/bastionhost-cn-2</a>******</para>
+        /// </summary>
+        [NameInMap("DownloadUrl")]
+        [Validation(Required=false)]
+        public string DownloadUrl { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>1679393152</para>
+        /// </summary>
+        [NameInMap("ExpireTime")]
+        [Validation(Required=false)]
+        public long? ExpireTime { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>2</para>
+        /// </summary>
+        [NameInMap("JobId")]
+        [Validation(Required=false)]
+        public string JobId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>message</para>
+        /// </summary>
+        [NameInMap("Message")]
+        [Validation(Required=false)]
+        public string Message { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>98DBE5C2-7D7A-5393-9E5A-71074336D33B</para>
+        /// </summary>
+        [NameInMap("RequestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>Success</para>
+        /// </summary>
+        [NameInMap("Status")]
+        [Validation(Required=false)]
+        public string Status { get; set; }
+
+    }
+
+}