+++ yundun-bastionhost-20191209/core/Models/GetInstanceStoreInfoResponse.cs
@@ -1,0 +1,26 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Yundun_bastionhost20191209.Models
+{
+    public class GetInstanceStoreInfoResponse : TeaModel {
+        [NameInMap("headers")]
+        [Validation(Required=false)]
+        public Dictionary<string, string> Headers { get; set; }
+
+        [NameInMap("statusCode")]
+        [Validation(Required=false)]
+        public int? StatusCode { get; set; }
+
+        [NameInMap("body")]
+        [Validation(Required=false)]
+        public GetInstanceStoreInfoResponseBody Body { get; set; }
+
+    }
+
+}