+++ yundun-bastionhost-20191209/core/Models/ListTagKeysRequest.cs
@@ -40,6 +40,10 @@ namespace AlibabaCloud.SDK.Yundun_bastionhost20191209.Models
         [Validation(Required=false)]
         public string RegionId { get; set; }
 
+        [NameInMap("ResourceGroupId")]
+        [Validation(Required=false)]
+        public string ResourceGroupId { get; set; }
+
         /// <summary>
         /// <para>The type of the resource.</para>
         /// <para>Set the value to INSTANCE, which indicates that the resource is a bastion host.</para>