+++ yundun-bastionhost-20191209/core/Models/ListTagResourcesRequest.cs
@@ -30,6 +30,10 @@ namespace AlibabaCloud.SDK.Yundun_bastionhost20191209.Models
         [Validation(Required=false)]
         public string RegionId { get; set; }
 
+        [NameInMap("ResourceGroupId")]
+        [Validation(Required=false)]
+        public string ResourceGroupId { get; set; }
+
         /// <summary>
         /// <para>The IDs of instances.</para>
         ///